+++ lib/rexx.php
@@ -1865,5 +1865,31 @@ class rexx extends rex {
 			return rexx::getString($key);
 		}
 	}
+
+	/**
+	 * Formats given bytes (e.g. from filesize()) to human readable string like 20,53 MB.
+	 *
+	 * @param int $bytes
+	 *
+	 * @return string
+	 *
+	 */
+	public static function formatBytes($bytes) {
+		if ($bytes >= 1073741824) {
+			$bytes = number_format($bytes / 1073741824, 2) . ' GB';
+		} elseif ($bytes >= 1048576) {
+			$bytes = number_format($bytes / 1048576, 2) . ' MB';
+		} elseif ($bytes >= 1024) {
+			$bytes = number_format($bytes / 1024, 2) . ' KB';
+		} elseif ($bytes > 1) {
+			$bytes = $bytes . ' bytes';
+		} elseif ($bytes == 1) {
+			$bytes = $bytes . ' byte';
+		} else {
+			$bytes = '0 bytes';
+		}
+
+		return $bytes;
+	}
 }