+++ CHANGELOG.md
@@ -7,6 +7,7 @@ X-Core - Changelog
 * X-Core Styles: Extra MForm Tabs Style entfernt
 * X-Core Styles: MForm Checkbox wieder ausgerichtet
 * rexx API: `rexx::sortArticles()` hinzugefügt, thx@fietstouring
+* rexx API: `rexx::formatBytes()` hinzugefügt, gibt Bytes z.b. von `filesize()` formatiert aus z.B. 20,53 MB.
 * rexx API: `rexx::threeWayComparison()` hinzugefügt, PHP5 Ersatz für Spaceship Operator `<=>` aus PHP 7
 * rexx API: `rexx::sanitizeFormValue()` hinzugefügt, bereinigt den Wert eines Formulars (Helfer für Custom Frontend Formular)
 * rexx API: `rexx::validateFormData()` hinzugefügt, validiert den Wert eines Formulars (Helfer für Custom Frontend Formular)